refactorings

+++ src/main/java/pixelitor/tools/brushes/IdealBrush.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2014 Laszlo Balazs-Csiki
+ * Copyright (c) 2015 Laszlo Balazs-Csiki
  *
  * This file is part of Pixelitor. Pixelitor is free software: you
  * can redistribute it and/or modify it under the terms of the GNU
@@ -8,12 +8,13 @@
  *
  * Pixelitor is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with Pixelitor.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Pixelitor. If not, see <http://www.gnu.org/licenses/>.
  */
+
 package pixelitor.tools.brushes;
 
 import pixelitor.tools.StrokeType;
@@ -32,6 +33,6 @@ public class IdealBrush extends StrokeBrush {
     @Override
     public void drawPoint(Graphics2D g, int x, int y, int radius) {
         int diameter = 2 * radius;
-        g.fillOval(x - radius, y - radius, (int) diameter, (int) diameter);
+        g.fillOval(x - radius, y - radius, diameter, diameter);
     }
 }
